[iOS] Extract method for creating strings from TimeInterval to string extensions.
